requirements.txt
pytest
requests
playwright
pytest-playwright
pyYaml